+++ src/main/strategies/AboveMiddleTaxStrategy.java
@@ -1,7 +1,7 @@
 package main.strategies;
 
 public class AboveMiddleTaxStrategy extends TaxStrategy {
-    public AboveMiddleTaxStrategy(Double grossIncome) {
+    public AboveMiddleTaxStrategy(long grossIncome) {
         super(grossIncome);
         this.setBasicTax(19822.0);
         this.setTaxPoint(87000.0);
